+++ validators/createFolderValidator.js
@@ -1,4 +1,9 @@
-const { checkSchema, validationResult } = require("express-validator");
+const asyncHandler = require("express-async-handler");
+const {
+  checkSchema,
+  validationResult,
+  matchedData,
+} = require("express-validator");
 
 const folderSchema = {
   folder_name: {
@@ -16,21 +21,22 @@ const folderSchema = {
   },
 };
 
-const validateFolder = checkSchema(folderSchema, ["body"]);
+// const validateFolder = checkSchema(folderSchema, ["body"]);
 
-const foo = (view) => {
-  return (req, res, next) => {
-    checkSchema(folderSchema, ["body"]);
+const validateFolder = (view) => {
+  return asyncHandler(async (req, res, next) => {
+    await checkSchema(folderSchema, ["body"]).run(req);
     const errors = validationResult(req);
     const inputs = matchedData(req, { onlyValidData: false });
-
+    console.log(errors);
     if (!errors.isEmpty()) {
       // Need a way to render the dashboard with the dialog open with createFolderForm partial
-      return res.render(view, {
-        formType: "folder",
+      return res.status(422).render(view, {
+        errors: errors.mapped(),
+        inputs,
       });
     }
-  };
+  });
 };
 
 module.exports = validateFolder;